refactor(engine): adding registerDecorators API
