+++ packages/lwc-engine/src/framework/main.ts
@@ -6,6 +6,7 @@ export { register } from "./services";
 export { unwrap } from "./membrane";
 export { registerTemplate } from "./template";
 export { registerComponent } from "./component";
+export { registerDecorators } from "./decorators/register";
 export { isNodeFromTemplate } from "./vm";
 
 // TODO: REMOVE THIS https://github.com/salesforce/lwc/issues/129
